refactor module organisation
